update dependencies and migrate to fork

+++ ptyutil/winsize.go
@@ -1,3 +1,4 @@
+//go:build !solaris
 // +build !solaris
 
 package ptyutil
@@ -15,12 +16,12 @@ type ttySize struct {
 	Ypixel uint16
 }
 
-func SetWinSize(f *os.File, cols int, rows int) error {
+func SetWinSize(f *os.File, cols uint16, rows uint16) error {
 	_, _, e := syscall.Syscall6(
 		syscall.SYS_IOCTL,
 		uintptr(f.Fd()),
 		uintptr(syscall.TIOCSWINSZ),
-		uintptr(unsafe.Pointer(&ttySize{uint16(rows), uint16(cols), 0, 0})),
+		uintptr(unsafe.Pointer(&ttySize{rows, cols, 0, 0})),
 		0, 0, 0,
 	)
